+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ftc16072/BehaviorTrees/Actions/Trajectories/MiddleSpikeTrajectory/ToBackboard/SetMiddleSpikeTrajectoryPart2.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.ftc16072.BehaviorTrees.Actions.Trajectories.LeftSpikeTrajectory;
+package org.firstinspires.ftc.teamcode.ftc16072.BehaviorTrees.Actions.Trajectories.MiddleSpikeTrajectory.ToBackboard;
 
 import org.firstinspires.ftc.teamcode.ftc16072.BehaviorTrees.Node;
 import org.firstinspires.ftc.teamcode.ftc16072.OpModes.QQOpMode;
@@ -6,24 +6,21 @@ import org.firstinspires.ftc.teamcode.ftc16072.Util.SpikePosition;
 import org.firstinspires.ftc.teamcode.rr_trajectorysequence.TrajectorySequence;
 import org.firstinspires.ftc.teamcode.rr_trajectorysequence.TrajectorySequenceBuilder;
 
-public class SetLeftSpikeTrajectoryPart2 extends Node {
-
-    public static final int LEFT_DISTANCE_INCHES = 8;
+public class SetMiddleSpikeTrajectoryPart2 extends Node {
+    public static final int FORWARD_DISTANCE_INCHES = 24;
+    public static final int BACKWARD_DISTANCE_INCHES = 20;
 
     @Override
     public State tick(QQOpMode opMode) {
         TrajectorySequenceBuilder builder = opMode.robot.nav.trajectorySequenceBuilder(opMode.robot.nav.getPoseEstimate());
         TrajectorySequence sequence = builder
-                //right because robot is facing backward
-                .strafeRight(LEFT_DISTANCE_INCHES)
-                .build();
+                .back(FORWARD_DISTANCE_INCHES).
+                build();
         if (sequence == null){
             return State.FAILURE;
         }
         opMode.board.add("TrajectorySequence", sequence);
-        opMode.board.add("TrajectorySequenceString", "Left Spike Trajectory 2");
-
-        opMode.board.add("SpikePosition", SpikePosition.LEFT);
+        opMode.board.add("SpikePosition", SpikePosition.MIDDLE);
         return State.SUCCESS;
     }
 }
