+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ftc16072/BehaviorTrees/Actions/Trajectories/RightSpikeTrajectory/Spike/SetRightSpikeTrajectory.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.ftc16072.BehaviorTrees.Actions.Trajectories;
+package org.firstinspires.ftc.teamcode.ftc16072.BehaviorTrees.Actions.Trajectories.RightSpikeTrajectory.Spike;
 
 import org.firstinspires.ftc.teamcode.ftc16072.BehaviorTrees.Node;
 import org.firstinspires.ftc.teamcode.ftc16072.OpModes.QQOpMode;
@@ -6,21 +6,20 @@ import org.firstinspires.ftc.teamcode.ftc16072.Util.SpikePosition;
 import org.firstinspires.ftc.teamcode.rr_trajectorysequence.TrajectorySequence;
 import org.firstinspires.ftc.teamcode.rr_trajectorysequence.TrajectorySequenceBuilder;
 
-public class SetMiddleSpikeTrajectory extends Node {
-    public static final int FORWARD_DISTANCE_INCHES = 24;
-    public static final int BACKWARD_DISTANCE_INCHES = 20;
+public class SetRightSpikeTrajectory extends Node {
+    public static final int FORWARD_DISTANCE_INCHES = 20;
+
 
     @Override
     public State tick(QQOpMode opMode) {
         TrajectorySequenceBuilder builder = opMode.robot.nav.trajectorySequenceBuilder(opMode.robot.nav.getPoseEstimate());
-        TrajectorySequence sequence = builder
-                .back(FORWARD_DISTANCE_INCHES).
-                build();
+        TrajectorySequence sequence = builder.back(FORWARD_DISTANCE_INCHES).
+                                            build();
         if (sequence == null){
             return State.FAILURE;
         }
         opMode.board.add("TrajectorySequence", sequence);
-        opMode.board.add("SpikePosition", SpikePosition.MIDDLE);
+        opMode.board.add("SpikePosition", SpikePosition.RIGHT);
         return State.SUCCESS;
     }
 }
