+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ftc16072/BehaviorTrees/Actions/Trajectories/RightSpikeTrajectory/Spike/SetRightSpikeTrajectoryPart2.java
@@ -1,0 +1,25 @@
+package org.firstinspires.ftc.teamcode.ftc16072.BehaviorTrees.Actions.Trajectories.RightSpikeTrajectory.Spike;
+
+import org.firstinspires.ftc.teamcode.ftc16072.BehaviorTrees.Node;
+import org.firstinspires.ftc.teamcode.ftc16072.OpModes.QQOpMode;
+import org.firstinspires.ftc.teamcode.ftc16072.Util.SpikePosition;
+import org.firstinspires.ftc.teamcode.rr_trajectorysequence.TrajectorySequence;
+import org.firstinspires.ftc.teamcode.rr_trajectorysequence.TrajectorySequenceBuilder;
+
+public class SetRightSpikeTrajectoryPart2 extends Node {
+
+    public static final int RIGHT_DISTANCE_INCHES = 20;
+
+    @Override
+    public State tick(QQOpMode opMode) {
+        TrajectorySequenceBuilder builder = opMode.robot.nav.trajectorySequenceBuilder(opMode.robot.nav.getPoseEstimate());
+        TrajectorySequence sequence = builder.strafeLeft(RIGHT_DISTANCE_INCHES)
+                        .build();
+        if (sequence == null){
+            return State.FAILURE;
+        }
+        opMode.board.add("TrajectorySequence", sequence);
+        opMode.board.add("SpikePosition", SpikePosition.RIGHT);
+        return State.SUCCESS;
+    }
+}
