+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ftc16072/BehaviorTrees/Actions/Trajectories/RightSpikeTrajectory/ToBackboard/SetRightSpikeTrajectoryPart3.java
@@ -1,0 +1,25 @@
+package org.firstinspires.ftc.teamcode.ftc16072.BehaviorTrees.Actions.Trajectories.RightSpikeTrajectory.ToBackboard;
+
+import org.firstinspires.ftc.teamcode.ftc16072.BehaviorTrees.Node;
+import org.firstinspires.ftc.teamcode.ftc16072.OpModes.QQOpMode;
+import org.firstinspires.ftc.teamcode.ftc16072.Util.SpikePosition;
+import org.firstinspires.ftc.teamcode.rr_trajectorysequence.TrajectorySequence;
+import org.firstinspires.ftc.teamcode.rr_trajectorysequence.TrajectorySequenceBuilder;
+
+public class SetRightSpikeTrajectoryPart3 extends Node {
+    //EDIT (as of 2/18)
+    public static final int FORWARD_DISTANCE_INCHES=20;
+
+    @Override
+    public State tick(QQOpMode opMode) {
+        TrajectorySequenceBuilder builder = opMode.robot.nav.trajectorySequenceBuilder(opMode.robot.nav.getPoseEstimate());
+        TrajectorySequence sequence = builder.back(FORWARD_DISTANCE_INCHES).
+                                            build();
+        if (sequence == null){
+            return State.FAILURE;
+        }
+        opMode.board.add("TrajectorySequence", sequence);
+        opMode.board.add("SpikePosition", SpikePosition.RIGHT);
+        return State.SUCCESS;
+    }
+}
